fix: share loader across modules
